+++ src/QueryBuilder/Condition/NotBetweenColumns.php
@@ -4,23 +4,15 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\QueryBuilder\Condition;
 
-use Iterator;
-use InvalidArgumentException;
-use Yiisoft\Db\Expression\ExpressionInterface;
-
-use function is_array;
-use function is_int;
-use function is_string;
-
 /**
- * Represents a `BETWEEN` operator where values are between two columns.
+ * Represents a `NOT BETWEEN` operator where values are between two columns.
  *
  * For example:
  *
  * ```php
  * new BetweenColumnsCondition(42, 'min_value', 'max_value')
  * // Will be build to:
- * // 42 BETWEEN min_value AND max_value
+ * // 42 NOT BETWEEN min_value AND max_value
  * ```
  *
  * And a more complex example:
@@ -33,9 +25,9 @@ use function is_string;
  * );
  *
  * // Will be built to:
- * // NOW() BETWEEN (SELECT time FROM log ORDER BY id ASC LIMIT 1) AND update_time
+ * // NOW() NOT BETWEEN (SELECT time FROM log ORDER BY id ASC LIMIT 1) AND update_time
  * ```
  */
-final class BetweenColumns extends AbstractBetweenColumns
+final class NotBetweenColumns extends AbstractBetweenColumns
 {
 }