+++ tests/Provider/QueryBuilderProvider.php
@@ -20,6 +20,7 @@ use Yiisoft\Db\QueryBuilder\Condition\BetweenColumns;
 use Yiisoft\Db\QueryBuilder\Condition\In;
 use Yiisoft\Db\QueryBuilder\Condition\Like;
 use Yiisoft\Db\QueryBuilder\Condition\LikeMode;
+use Yiisoft\Db\QueryBuilder\Condition\NotBetweenColumns;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 use Yiisoft\Db\Schema\Column\ColumnBuilder;
 use Yiisoft\Db\Schema\Data\StringableStream;
@@ -344,29 +345,28 @@ class QueryBuilderProvider
                 [':qp0' => 123],
             ],
             [
-                new BetweenColumns('2018-02-11', 'BETWEEN', 'create_time', 'update_time'),
+                new BetweenColumns('2018-02-11', 'create_time', 'update_time'),
                 ':qp0 BETWEEN [[create_time]] AND [[update_time]]',
                 [':qp0' => '2018-02-11'],
             ],
             [
-                new BetweenColumns('2018-02-11', 'NOT BETWEEN', 'NOW()', 'update_time'),
+                new NotBetweenColumns('2018-02-11', 'NOW()', 'update_time'),
                 ':qp0 NOT BETWEEN NOW() AND [[update_time]]',
                 [':qp0' => '2018-02-11'],
             ],
             [
-                new BetweenColumns(new Expression('NOW()'), 'BETWEEN', 'create_time', 'update_time'),
+                new BetweenColumns(new Expression('NOW()'), 'create_time', 'update_time'),
                 'NOW() BETWEEN [[create_time]] AND [[update_time]]',
                 [],
             ],
             [
-                new BetweenColumns(new Expression('NOW()'), 'NOT BETWEEN', 'create_time', 'update_time'),
+                new NotBetweenColumns(new Expression('NOW()'), 'create_time', 'update_time'),
                 'NOW() NOT BETWEEN [[create_time]] AND [[update_time]]',
                 [],
             ],
             [
-                new BetweenColumns(
+                new NotBetweenColumns(
                     new Expression('NOW()'),
-                    'NOT BETWEEN',
                     (new Query(static::getDb()))->select('min_date')->from('some_table'),
                     'max_date'
                 ),
@@ -374,9 +374,8 @@ class QueryBuilderProvider
                 [],
             ],
             [
-                new BetweenColumns(
+                new NotBetweenColumns(
                     new Expression('NOW()'),
-                    'NOT BETWEEN',
                     new Expression('min_date'),
                     (new Query(static::getDb()))->select('max_date')->from('some_table'),
                 ),